generic attributes are allowed in C# 11
The note in the generics and attributes article said that generic classes can't derive from `System.Attribute`. That changes in C# 11.

While making that change, move snippets into this folder.
