+++ docs/csharp/programming-guide/generics/snippets/GenericWhereConstraints.cs
@@ -1,8 +1,4 @@
-﻿using System;
-using System.Collections;
-using System.Collections.Generic;
-using System.Text;
-
+﻿
 using static generics.UnmanagedExtensions;
 
 namespace generics
