+++ docs/csharp/programming-guide/generics/snippets/Program.cs
@@ -1,14 +1,2 @@
-﻿using System;
-using System.Threading.Tasks;
-
-namespace generics
-{
-    class Program
-    {
-        static void Main(string[] args)
-        {
-            Console.WriteLine("=================    Generic Where Constraints Examples ======================");
-            GenericWhereConstraints.Examples();
-        }
-    }
-}
+﻿Console.WriteLine("=================    Generic Where Constraints Examples ======================");
+generics.GenericWhereConstraints.Examples();
