+++ docs/csharp/programming-guide/generics/snippets/generics.csproj
@@ -2,8 +2,11 @@
 
   <PropertyGroup>
     <OutputType>Exe</OutputType>
-    <TargetFramework>netcoreapp3.1</TargetFramework>
+    <TargetFramework>net6.0</TargetFramework>
+    <Nullable>enable</Nullable>
+    <ImplicitUsings>enable</ImplicitUsings>
     <AllowUnsafeBlocks>true</AllowUnsafeBlocks>
     <StartupObject>generics.Program</StartupObject>
+    <LangVersion>preview</LangVersion>
   </PropertyGroup>
 </Project>
