+++ samples/snippets/csharp/VS_Snippets_VBCSharp/csProgGuideGenerics/CS/Generics.cs
@@ -703,51 +703,4 @@ namespace CsCsrefProgrammingGenerics
             //</Snippet46>
         }
     }
-
-    //---------------------------------------------------------------------------
-    class WrapAttributes
-    {
-        //-----------------------------------------------------
-        //<Snippet48>
-        class CustomAttribute : System.Attribute
-        {
-            public System.Object info;
-        }
-        //</Snippet48>
-
-        //-----------------------------------------------------
-        //<Snippet49>
-        public class GenericClass1<T> { }
-
-        [CustomAttribute(info = typeof(GenericClass1<>))]
-        class ClassA { }
-        //</Snippet49>
-
-        //-----------------------------------------------------
-        //<Snippet50>
-        public class GenericClass2<T, U> { }
-
-        [CustomAttribute(info = typeof(GenericClass2<,>))]
-        class ClassB { }
-        //</Snippet50>
-
-        //-----------------------------------------------------
-        //<Snippet51>
-        public class GenericClass3<T, U, V> { }
-
-        [CustomAttribute(info = typeof(GenericClass3<int, double, string>))]
-        class ClassC { }
-        //</Snippet51>
-
-        //-----------------------------------------------------
-        //<Snippet52>
-        //[CustomAttribute(info = typeof(GenericClass3<int, T, string>))]  //Error
-        class ClassD<T> { }
-        //</Snippet52>
-
-        //-----------------------------------------------------
-        //<Snippet53>
-        //public class CustomAtt<T> : System.Attribute {}  //Error
-        //</Snippet53>
-    }
 }
